More accurate range values and some cleaning up
For some reason I have a bunch of if thens to find equipped weapon info, followed by a switch that checked for a var that I set within the if thens to change other stuff. I have no idea why, that seems horribly inefficient. So I got rid of that switch entirely.
Also range values should be hopefully be much more accurate